+++ source/js/winning-scene/easing.js
@@ -1,0 +1,55 @@
+export function easeLinear(x) {
+  return x;
+}
+
+export function easeInCubic(x) {
+  return x ** 3;
+}
+
+export function easeInQuad(x) {
+  return x ** 4;
+}
+
+export function easeOutCubic(x) {
+  return 1 - Math.pow(1 - x, 3);
+}
+
+export function easeInExpo(x) {
+  if (x === 0) {
+    return 0;
+  } else {
+    return Math.pow(2, 10 * x - 10);
+  }
+}
+
+export function easeOutExpo(x) {
+  if (x === 1) {
+    return 1;
+  } else {
+    return 1 - Math.pow(2, -10 * x);
+  }
+}
+
+export function easeInElastic(x) {
+  const c4 = (2 * Math.PI) / 3;
+
+  if (x === 0) {
+    return 0;
+  } else if (x === 1) {
+    return 1;
+  } else {
+    return Math.pow(2, 10 * x - 10) * Math.sin((x * 10 - 10.75) * c4);
+  }
+}
+
+export function easeOutElastic(x) {
+  const c4 = (2 * Math.PI) / 3;
+
+  if (x === 0) {
+    return 0;
+  } else if (x === 1) {
+    return 1;
+  } else {
+    return Math.pow(2, -10 * x) * Math.sin((x * 10 - 0.75) * c4) + 1;
+  }
+}
